Move timeline transform into standalone component
This is in preparation for adding more logic to timeline transform
